# openapi.json
{"openapi":"3.1.0","info":{"title":"Loyalty Integration API","description":"外部整合層，串接 Laravel Multi-Tenant Loyalty Platform","version":"1.0.0"},"paths":{"/auth/login":{"post":{"tags":["auth"],"summary":"Auth Login","description":"手動登入取得 JWT Token","operationId":"auth_login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/me":{"get":{"tags":["auth"],"summary":"Auth Me","description":"取得目前登入的使用者資訊","operationId":"auth_me_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/refresh":{"post":{"tags":["auth"],"summary":"Auth Refresh","description":"刷新 JWT Token","operationId":"auth_refresh_auth_refresh_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/logout":{"post":{"tags":["auth"],"summary":"Auth Logout","description":"登出並清除目前的 Token","operationId":"auth_logout_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/customers":{"get":{"tags":["customers"],"summary":"List Customers","operationId":"list_customers_customers_get","parameters":[{"name":"per_page","in":"query","required":false,"schema":{"type":"integer","default":15,"title":"Per Page"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/customers/{customer_id}":{"get":{"tags":["customers"],"summary":"Get Customer","operationId":"get_customer_customers__customer_id__get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"integer","title":"Customer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/customers/{customer_id}/qr-code":{"get":{"tags":["customers"],"summary":"Get Qr Code","description":"取得會員 QR Code（回傳 base64 SVG）","operationId":"get_qr_code_customers__customer_id__qr_code_get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"integer","title":"Customer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/customers/{customer_id}/membership":{"get":{"tags":["customers"],"summary":"Get Membership","description":"取得會員等級資訊","operationId":"get_membership_customers__customer_id__membership_get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"integer","title":"Customer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/customers/identify":{"post":{"tags":["customers"],"summary":"Identify By Qr","description":"透過 QR Token 識別會員（模擬 POS 掃碼）","operationId":"identify_by_qr_customers_identify_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdentifyByQrRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/customers/{customer_id}/points":{"get":{"tags":["points"],"summary":"Get Balance","operationId":"get_balance_customers__customer_id__points_get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"integer","title":"Customer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/customers/{customer_id}/point-transactions":{"get":{"tags":["points"],"summary":"List Point Transactions","operationId":"list_point_transactions_customers__customer_id__point_transactions_get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"integer","title":"Customer Id"}},{"name":"per_page","in":"query","required":false,"schema":{"type":"integer","default":15,"title":"Per Page"}},{"name":"type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["points"],"summary":"Create Point Transaction","description":"建立通用點數交易 - 對應 Laravel POST /api/v1/customers/{customer}/point-transactions","operationId":"create_point_transaction_customers__customer_id__point_transactions_post","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"integer","title":"Customer Id"}},{"name":"Idempotency-Key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PointTransactionRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/customers/{customer_id}/point-transactions/expiring":{"get":{"tags":["points"],"summary":"Get Expiring Transactions","operationId":"get_expiring_transactions_customers__customer_id__point_transactions_expiring_get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"integer","title":"Customer Id"}},{"name":"days","in":"query","required":false,"schema":{"type":"integer","default":30,"title":"Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/customers/{customer_id}/point-transactions/{transaction_id}":{"get":{"tags":["points"],"summary":"Get Point Transaction","operationId":"get_point_transaction_customers__customer_id__point_transactions__transaction_id__get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"integer","title":"Customer Id"}},{"name":"transaction_id","in":"path","required":true,"schema":{"type":"integer","title":"Transaction Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/points/earn":{"post":{"tags":["points"],"summary":"Earn Points","operationId":"earn_points_points_earn_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EarnRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/points/redeem":{"post":{"tags":["points"],"summary":"Redeem Points","operationId":"redeem_points_points_redeem_post","parameters":[{"name":"Idempotency-Key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RedeemRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/customers/{customer_id}/coupons":{"get":{"tags":["coupons"],"summary":"List Coupons","operationId":"list_coupons_customers__customer_id__coupons_get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"integer","title":"Customer Id"}},{"name":"per_page","in":"query","required":false,"schema":{"type":"integer","default":15,"title":"Per Page"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/customers/{customer_id}/coupons/{user_coupon_id}":{"get":{"tags":["coupons"],"summary":"Get Coupon","description":"取得單一會員優惠券的詳細資訊","operationId":"get_coupon_customers__customer_id__coupons__user_coupon_id__get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"integer","title":"Customer Id"}},{"name":"user_coupon_id","in":"path","required":true,"schema":{"type":"integer","title":"User Coupon Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/customers/{customer_id}/coupon-redemptions":{"get":{"tags":["coupons"],"summary":"List Coupon Redemptions","description":"取得會員的優惠券核銷歷史記錄","operationId":"list_coupon_redemptions_customers__customer_id__coupon_redemptions_get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"integer","title":"Customer Id"}},{"name":"per_page","in":"query","required":false,"schema":{"type":"integer","default":15,"title":"Per Page"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/coupons/redeem":{"post":{"tags":["coupons"],"summary":"Redeem Coupon","operationId":"redeem_coupon_coupons_redeem_post","parameters":[{"name":"Idempotency-Key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CouponRedeemRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/payments/mixed":{"post":{"tags":["coupons"],"summary":"Mixed Payment","operationId":"mixed_payment_payments_mixed_post","parameters":[{"name":"Idempotency-Key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MixedPaymentRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/customers/{customer_id}/coupons/claim":{"post":{"tags":["coupons"],"summary":"Claim Coupon","description":"領取優惠券 - 對應 Laravel POST /api/v1/customers/{customer}/coupons/claim","operationId":"claim_coupon_customers__customer_id__coupons_claim_post","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"integer","title":"Customer Id"}},{"name":"Idempotency-Key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CouponClaimRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/coupons/claim":{"post":{"tags":["coupons"],"summary":"Claim Coupon Via Api","description":"領取優惠券 - 統一 API 格式","operationId":"claim_coupon_via_api_coupons_claim_post","parameters":[{"name":"Idempotency-Key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CouponClaimRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/customers/{customer_id}/reward-grants":{"get":{"tags":["rewards"],"summary":"List Reward Grants","description":"列出客戶的所有獎勵發放記錄 - 對應 Laravel GET /api/v1/customers/{customer}/reward-grants","operationId":"list_reward_grants_customers__customer_id__reward_grants_get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"integer","title":"Customer Id"}},{"name":"per_page","in":"query","required":false,"schema":{"type":"integer","default":15,"title":"Per Page"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/customers/{customer_id}/rewards/grant":{"post":{"tags":["rewards"],"summary":"Grant Reward","description":"發放獎勵給客戶 - 對應 Laravel POST /api/v1/customers/{customer}/rewards/grant","operationId":"grant_reward_customers__customer_id__rewards_grant_post","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"integer","title":"Customer Id"}},{"name":"Idempotency-Key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GrantRewardRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/workflows/pos-checkout":{"post":{"tags":["workflows"],"summary":"Pos Checkout","operationId":"pos_checkout_workflows_pos_checkout_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/POSCheckoutRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"CouponClaimRequest":{"properties":{"customer_id":{"type":"integer","title":"Customer Id"},"code":{"type":"string","title":"Code","description":"優惠券代碼"}},"type":"object","required":["customer_id","code"],"title":"CouponClaimRequest"},"CouponRedeemRequest":{"properties":{"customer_id":{"type":"integer","title":"Customer Id"},"user_coupon_id":{"type":"integer","title":"User Coupon Id"},"reference":{"type":"string","title":"Reference"},"order_reference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Order Reference"},"order_amount":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Order Amount"}},"type":"object","required":["customer_id","user_coupon_id","reference"],"title":"CouponRedeemRequest"},"EarnRequest":{"properties":{"customer_id":{"type":"integer","title":"Customer Id"},"amount":{"type":"integer","exclusiveMinimum":0.0,"title":"Amount"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"reference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reference"}},"type":"object","required":["customer_id","amount"],"title":"EarnRequest"},"GrantRewardRequest":{"properties":{"campaign_reward_id":{"type":"integer","title":"Campaign Reward Id"},"quantity":{"type":"integer","title":"Quantity","default":1},"reference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reference"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["campaign_reward_id"],"title":"GrantRewardRequest","description":"獎勵發放請求 - 對應 Laravel POST /api/v1/customers/{customer}/rewards/grant"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IdentifyByQrRequest":{"properties":{"qr_token":{"type":"string","title":"Qr Token"}},"type":"object","required":["qr_token"],"title":"IdentifyByQrRequest"},"LoginRequest":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"MixedPaymentRequest":{"properties":{"customer_id":{"type":"integer","title":"Customer Id"},"reference":{"type":"string","title":"Reference"},"order_amount":{"type":"integer","minimum":0.0,"title":"Order Amount"},"user_coupon_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"User Coupon Id"},"points_amount":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Points Amount"},"order_reference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Order Reference"}},"type":"object","required":["customer_id","reference","order_amount"],"title":"MixedPaymentRequest"},"POSCheckoutRequest":{"properties":{"customer_id":{"type":"integer","title":"Customer Id"},"earn_amount":{"type":"integer","minimum":0.0,"title":"Earn Amount","default":100},"redeem_amount":{"type":"integer","minimum":0.0,"title":"Redeem Amount","default":0},"order_reference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Order Reference"}},"type":"object","required":["customer_id"],"title":"POSCheckoutRequest"},"PointTransactionRequest":{"properties":{"type":{"type":"string","title":"Type","description":"交易類型: earn, redeem, adjust, etc."},"amount":{"type":"integer","exclusiveMinimum":0.0,"title":"Amount","description":"交易金額（必須為正數）"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"交易描述"},"reference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reference","description":"參考編號"}},"type":"object","required":["type","amount"],"title":"PointTransactionRequest","description":"通用點數交易建立請求 - 對應 Laravel POST /api/v1/customers/{customer}/point-transactions"},"RedeemRequest":{"properties":{"customer_id":{"type":"integer","title":"Customer Id"},"amount":{"type":"integer","exclusiveMinimum":0.0,"title":"Amount"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","default":"Redeemed via Integration API"},"reference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reference"}},"type":"object","required":["customer_id","amount"],"title":"RedeemRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}